requirements.txt
Django
django-browser-reload
djangorestframework
djangorestframework-simplejwt
django-cors-headers